% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_basemap.R
\name{get_basemap}
\alias{get_basemap}
\title{Download IPBES World Countries Boundaries}
\usage{
get_basemap(path = here::here("data", "ipbes-regions"), force = FALSE)
}
\arguments{
\item{path}{the folder to download and extract zip file}

\item{force}{a logical value. If \code{TRUE} will erase previously downloaded
files. Default is \code{FALSE}.}
}
\description{
This function downloads shapefile of world countries boundaries as defined by
IPBES (available at \url{https://zenodo.org/record/3928281}). Zip file is
downloaded in the folder \code{path} ('data/ipbes-regions' by default) and files
are extracted in the same folder. If this folder doesn't exist it will be
created.

\strong{Note:} if you delete the ZIP file (after files extraction) and re-run this
function, the ZIP file will be downloaded again. It this is what you want do
forget to use \code{force = TRUE} to erase previous files.

Original Projection System: WGS84
}
\examples{
\dontrun{
## Download and extract ZIP ----
get_basemap()

## Check ----
list.files(here::here("data", "ipbes-regions"))

## Re-run function ----
get_basemap()              # No download (ZIP already present)
get_basemap(force = TRUE)  # Erase previous files
}
}
