% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worldmap-package.R
\docType{package}
\name{worldmap-package}
\alias{worldmap}
\alias{worldmap-package}
\title{worldmap: The Title of the Project}
\description{
A paragraph providing a full description of the project (on
    several lines...)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/frbcesab/worldmap}
  \item Report bugs at \url{https://github.com/frbcesab/worldmap/issues}
}

}
\author{
\strong{Maintainer}: Nicolas Casajus \email{nicolas.casajus@gmail.com} (\href{https://orcid.org/0000-0002-5537-5294}{ORCID})

}
\keyword{internal}
